refactor(next): address welcome screen review feedback
Split LiveDemo into BlockPanel and WritePanel, centralize breakpoints and popover primitives, remove acc prop drilling, and dedupe dismiss and address formatting helpers.
