+++ templates/next/components/welcome/format.ts
@@ -1,0 +1,5 @@
+/** Truncate hex addresses and hashes for compact UI labels. */
+export function formatAddress(value: string, head = 6, tail = 4): string {
+  if (value.length <= head + tail + 1) return value;
+  return `${value.slice(0, head)}…${value.slice(-tail)}`;
+}
